Move model classes into Model package
