+++ app/src/main/java/com/example/android/climbtogether/Model/Gym.java
@@ -1,4 +1,6 @@
-package com.example.android.climbtogether;
+package com.example.android.climbtogether.Model;
+
+import com.example.android.climbtogether.Model.Problem;
 
 /**
  * Created by MD on 2017-01-28.
