+++ app/src/main/java/com/example/android/climbtogether/fragment/GymFragment.java
@@ -10,7 +10,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import com.example.android.climbtogether.Gym;
+import com.example.android.climbtogether.Model.Gym;
 import com.example.android.climbtogether.GymAdapter;
 import com.example.android.climbtogether.R;
 import com.example.android.climbtogether.activity.GymDetailActivity;
