+++ app/src/main/java/com/example/android/climbtogether/fragment/ProblemFragment.java
@@ -3,25 +3,21 @@ package com.example.android.climbtogether.fragment;
 import android.content.Intent;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ListView;
 
-import com.example.android.climbtogether.Problem;
+import com.example.android.climbtogether.Model.Problem;
 import com.example.android.climbtogether.ProblemAdapter;
 import com.example.android.climbtogether.R;
 import com.example.android.climbtogether.activity.ProblemDetailActivity;
 import com.example.android.climbtogether.activity.ProblemResister;
 import com.example.android.climbtogether.viewHolder.ProblemViewHolder;
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
-import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
@@ -29,8 +25,6 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
-
-import java.util.ArrayList;
 
 public class ProblemFragment extends Fragment {
