+++ app/src/main/java/com/example/android/climbtogether/viewHolder/ProblemViewHolder.java
@@ -10,7 +10,7 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.drawable.GlideDrawable;
 import com.bumptech.glide.request.RequestListener;
 import com.bumptech.glide.request.target.Target;
-import com.example.android.climbtogether.Problem;
+import com.example.android.climbtogether.Model.Problem;
 import com.example.android.climbtogether.R;
 
 /**
