+++ app/src/main/java/com/example/android/climbtogether/viewHolder/ViewHolder.java
@@ -10,7 +10,7 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.resource.drawable.GlideDrawable;
 import com.bumptech.glide.request.RequestListener;
 import com.bumptech.glide.request.target.Target;
-import com.example.android.climbtogether.Gym;
+import com.example.android.climbtogether.Model.Gym;
 import com.example.android.climbtogether.R;
 /**
  *    Created by MD on 2017-07-30.
